+++ bifast-outbound/src/main/java/bifast/outbound/corebank/DebitReversalSedaRoute.java
@@ -1,5 +1,6 @@
 package bifast.outbound.corebank;
 
+
 import org.apache.camel.Exchange;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
@@ -9,8 +10,10 @@ import org.springframework.stereotype.Component;
 
 import bifast.outbound.config.Config;
 import bifast.outbound.corebank.pojo.DebitReversalRequestPojo;
+import bifast.outbound.corebank.pojo.DebitReversalResponsePojo;
 import bifast.outbound.corebank.processor.DebitReversalFaultProcessor;
 import bifast.outbound.corebank.processor.DebitReversalRequestProcessor;
+import bifast.outbound.corebank.processor.SaveCbDebitReversalProc;
 import bifast.outbound.processor.EnrichmentAggregator;
 import bifast.outbound.service.JacksonDataFormatService;
 
@@ -21,11 +24,15 @@ public class DebitReversalSedaRoute extends RouteBuilder{
 	@Autowired private DebitReversalRequestProcessor debitReversalRequestProcessor;
 	@Autowired private EnrichmentAggregator enrichmentAggregator;
 	@Autowired private JacksonDataFormatService jdfService;
+	@Autowired private SaveCbDebitReversalProc saveCbDebitReversalProc;
+
+//	private static Logger logger = LoggerFactory.getLogger(DebitReversalSedaRoute.class);
 
 	@Override
 	public void configure() throws Exception {
 		
 		JacksonDataFormat debitReversalRequestJDF = jdfService.basic(DebitReversalRequestPojo.class);
+		JacksonDataFormat debitReversalResponseJDF = jdfService.wrapRoot(DebitReversalResponsePojo.class);
 
 		onException(Exception.class)
 			.handled(true)
@@ -38,7 +45,7 @@ public class DebitReversalSedaRoute extends RouteBuilder{
 					.log(LoggingLevel.ERROR, "${exception.stacktrace}")
 			.end()
 	    	.process(dbrevFaultProcessor)
-			.to("seda:savecbdebitrevr?exchangePattern=InOnly")
+			.process(saveCbDebitReversalProc)
 			;
 
 		from("seda:sdebitreversal").routeId("komi.cb.debit_rev")
@@ -51,24 +58,23 @@ public class DebitReversalSedaRoute extends RouteBuilder{
 			.log("[${exchangeProperty.prop_request_list.msgName}:${exchangeProperty.prop_request_list.requestId}] Request ISOAdapter: ${body}")
 
             .setHeader(Exchange.CONTENT_TYPE, constant("application/json"))
-
 			.setHeader("HttpMethod", constant("POST"))
 		
-			.enrich()
-				.simple("{{komi.url.isoadapter.reversal}}?bridgeEndpoint=true")
+			.enrich().simple("{{komi.url.isoadapter.reversal}}?"
+//					+ "socketTimeout=7000&" 
+					+ "bridgeEndpoint=true")
 				.aggregationStrategy(enrichmentAggregator)
-
 			.convertBodyTo(String.class)
 			
-			.log("[${exchangeProperty.prop_request_list.msgName}:${exchangeProperty.prop_request_list.requestId}] Response ISOAdapter: ${body}")
+			.log("[${header.cihubMsgName}:${exchangeProperty.prop_request_list.requestId}] Response ISOAdapter: ${body}")
+			.unmarshal(debitReversalResponseJDF)
+			
+//			.bean("SaveCbTrnsService", "debitReversal")
+			.process(saveCbDebitReversalProc)
 
-			.to("seda:savecbdebitrevr?exchangePattern=InOnly")
-		;
-		
-		
+			;
 		
 	}
 
-		
 		
 }
